requirements.txt
pandas
scikit-learn
xgboost
matplotlib
sentence-transformers
fastapi[standard]
langchain
langgraph
google-genai
joblib
pinecone
langchain-text-splitters
python-dotenv
langchain-google-genai
langchain-pinecone
langchain-huggingface
sqlalchemy 
psycopg2-binary 
alembic
pyjwt
pwdlib[argon2]
langchain-groq
redis
